% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_bibd.R
\name{anova_bibd}
\alias{anova_bibd}
\title{anova_bibd}
\usage{
anova_bibd(dados, x, y, bloco)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos}
}
\value{
list
}
\description{
Calcula valores utilizados para uma tabela de ANOVA de um delineamento em blocos incompletos balanceados
}
