% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anova_glsqd}
\alias{anova_glsqd}
\title{Gera valores de anova para Quadrados Latinos}
\usage{
anova_glsqd(dados, x, chi, y, linha, coluna)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados}

\item{x}{String com o nome da coluna dos tratamentos}

\item{chi}{String com o nome da coluna dos outros tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{linha}{String com o nome da coluna do data frame que representa a linha do quadrado latino}

\item{coluna}{String com o nome da coluna do data frame que representa a coluna do quadrado latino}
}
\value{
Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.
}
\description{
Calcula valores utilizados para uma tabela de ANOVA de um delineamento em quadrados latinos
}
\examples{

quadrados_latinos(4)$dados \%>\%
 anova_lsqd("tratamento", "resultado", "linha", "coluna")

}
