% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_lsqd.R
\name{anova_lsqd}
\alias{anova_lsqd}
\title{anova_lsqd}
\usage{
anova_lsqd(dados, x, y, linha, coluna)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{linha}{String com o nome da coluna do data frame que representa a linha do quadrado latino}

\item{coluna}{String com o nome da coluna do data frame que representa a coluna do quadrado latino}
}
\value{
list
}
\description{
Calcula valores utilizados para uma tabela de ANOVA de um delineamento em quadrados latinos
}
