% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{anova_rbcd}
\alias{anova_rbcd}
\title{Gera valores de anova para RBCD teste}
\usage{
anova_rbcd(dados, x, y, bloco)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos}
}
\value{
Objeto de tipo lista contendo todos os valores utilizados para o calculo da tabela da ANOVA, os valores da tabela da ANOVA e os parâmetros estimados.
}
\description{
Calcula valores utilizados para uma tabela de ANOVA de um delineamento em blocos completamente casualizados
}
\examples{

anova_rbcd(mtcars, "gear", "hp", "carb")

}
