% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloco_casualizado.R
\name{bloco_casualizado}
\alias{bloco_casualizado}
\title{Simula delineamento de blocos incompletos}
\usage{
bloco_casualizado(ntrat, nbloco, nrep = 1, fun = rnorm)
}
\arguments{
\item{ntrat}{Numero de tratamentos a serem simulados}

\item{nbloco}{Numero de blocos a serem simulados}

\item{fun}{Função que gere valores aleatórios cujo primeiro argumento é a quantidade de valores a serem gerados, Padrão = rnorm()}
}
\value{
Objeto de tipo lista com os elementos:
\itemize{
\item \strong{dados} Dados simulados em formato longo
\item \strong{dados_matriz} Dados simulados em formato largo, como matiz
\item \strong{kable} Texto em LaTeX da tabela pré-formatada
\item \strong{matriz} Matriz dos dados simulados
}
}
\description{
Simula um delineamento em blocos incompletos a partir de numeros de tratamentos, blocos e repetições.
}
\examples{

bloco_casualizado(4, 2)

}
