% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloco_incompleto.R
\name{bloco_incompleto}
\alias{bloco_incompleto}
\title{Simula delineamento de blocos incompletos}
\usage{
bloco_incompleto(ntrat, nbloco, nrep = NULL)
}
\arguments{
\item{ntrat}{Numero de tratamentos a serem simulados}

\item{nbloco}{Numero de blocos a serem simulados}

\item{nrep}{Numero de repeticoes a serem simuladas, caso haja, padrao = NULL}
}
\value{
Objeto de tipo lista com os elementos:
\itemize{
\item \strong{dados} Dados simulados em formato longo
\item \strong{dados_matriz} Dados simulados em formato largo, como matiz
\item \strong{kable} Texto em LaTeX da tabela pré-formatada
\item \strong{matriz} Matriz dos dados simulados
}
}
\description{
Simula um delineamento em blocos incompletos a partir de numeros de tratamentos, blocos e repetições.
}
\examples{

bloco_incompleto(8, 14, 3)$dados_matriz \%>\% kable_()

bloco_incompleto(4, 6, 2)

}
