% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparacoes.R
\name{comparacoes}
\alias{comparacoes}
\title{Comparacoes}
\usage{
comparacoes(x, y, bloco = NULL, dados, alpha = 0.05, group = FALSE)
}
\arguments{
\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos, manter NULL caso não hajam blocos}

\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos}

\item{alpha}{Nivel de significancia a ser utilizado, padrao 0.05}

\item{group}{Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos}
}
\value{
comparacao
}
\description{
Análises múltiplas para ANOVA de delineamentos inteiramente casualizados (DIC) e para delineamentos em blocos completos casualizados
}
