% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{kable_}
\alias{kable_}
\title{Kable pre configurada}
\usage{
kable_(x)
}
\arguments{
\item{x}{Data frame a ser printado}
}
\value{
Kable com colunas centralizadas
}
\description{
Kable modificada para centralizar todas as colunas
}
\examples{

kable_(mtcars)

}
