% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmult}
\alias{mmult}
\title{Multiplicação de matrizes}
\usage{
mmult(m, v)
}
\arguments{
\item{m}{Primeira matriz}

\item{v}{Segunda matriz}
}
\value{
Matriz.
}
\description{
Multplica duas matrizes.
}
