% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.comparacao}
\alias{plot.comparacao}
\title{Grafico de comparacoes (DIC, RBCD)}
\usage{
\method{plot}{comparacao}(comparacao, trat = TRUE, ...)
}
\arguments{
\item{comparacao}{Objeto de comparacao resultado da funcao \code{comparacao}}

\item{trat}{Variavel logica, caso verdadeira retorna somente o grafico de comparacoes dos tratamentos}
}
\value{
Ggplot com comparações
}
\description{
Graficos de comparacoes multiplas de delineamentos inteiramente casualizados e delineamentos em blocos complettamente casualizados
}
\examples{

plot(comparacoes("gear", "hp", "carb", mtcars))

}
