% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.comparacao_bib}
\alias{plot.comparacao_bib}
\title{Grafico de comparacoes (BIBD)}
\usage{
\method{plot}{comparacao_bib}(comparacao, trat = FALSE, ...)
}
\arguments{
\item{comparacao}{Objeto de comparacao_bib resultado da funcao `comparacao_bib`}

\item{trat}{Variavel logica, caso verdadeira retorna somente o grafico de comparacoes dos tratamentos}
}
\value{
ggplot
}
\description{
Graficos de comparacoes multiplas de delineamentos em blocos incompletos balanceados
}
