% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressupostos_anova.R
\name{pressupostos_anova}
\alias{pressupostos_anova}
\title{Análise de pressupostos (ANOVA)}
\usage{
pressupostos_anova(dados, x, y)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, linhas, colunas e, caso haja, blocos}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}
}
\value{
Lista que contém:
\itemize{
\item \code{dados} Dados alimentados a função.
\item \code{var}Nomes das variáveis utilizadas.
\item \code{formula} String da formula utilizada nos modelos.
\item \code{modelo} Modelo ajustado.
\item \code{num} Lista com valores utilizados para os calculos.
\item \code{teste} Lista com os diferentes testes utilizados.
\item \code{plots} Lista com os gráficos separados.
\item \code{plot} Ggplot com todos os gráficos unidos.
}
}
\description{
Análises de pressupostos para o uso da ANOVA
}
\examples{

bloco_casualizado(4, 4, 5)$dados \%>\%
  pressupostos_anova("Trat", "resultado")
 
}
