% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{print.pressupostos_anova}
\alias{print.pressupostos_anova}
\title{Print pressupostos}
\usage{
\method{print}{pressupostos_anova}(p)
}
\arguments{
\item{p}{Objeto de pressupostos}
}
\description{
Printa objeto de pressupostos de anova
}
