% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{print_anova}
\alias{print_anova}
\title{Print Anova}
\usage{
print_anova(model)
}
\arguments{
\item{model}{Modelo de anova}
}
\value{
Kable
}
\description{
Organiza e printa uma tabela de anova em formato de LaTeX
}
\examples{

aov(hp ~ cyl, data = mtcars) \%>\% print_anova()

}
