% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrados_glatinos.R
\name{quadrados_glatinos}
\alias{quadrados_glatinos}
\title{Simula delineamento de quadrados greco-latinos}
\usage{
quadrados_glatinos(n, fun = rnorm)
}
\arguments{
\item{n}{Número de linhas (ou colunas)}

\item{fun}{Função que gere valores aleatórios cujo primeiro argumento é a quantidade de valores a serem gerados, Padrão = rnorm()}

\item{nrep}{Número de réplicas a serem feitas, Padrão = 1}
}
\value{
Banco de dados com colunas dos tratamentos, linhas e colunas
}
\description{
Simula um delineamento em quadrados greco-latinos
}
\examples{

quadrados_glatinos(4)

}
