% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{quadrados_latinos}
\alias{quadrados_latinos}
\title{Quadrados latinos}
\usage{
quadrados_latinos(dados, tratamentos)
}
\arguments{
\item{dados}{Data frame com os resultados}

\item{tratamentos}{Vetor com fatores dos tratamentos}
}
\value{
data.frame
}
\description{
Simula um delineamento em quadrados latinos a partir de um banco de dados
}
