% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_sort}
\alias{Rcpp_sort}
\title{Ordenar vetor}
\usage{
Rcpp_sort(x)
}
\arguments{
\item{x}{Vetor}
}
\value{
Vetor numerico com indices ordenados de forma que os dados fiquem ordenados.
}
\description{
Gera um vetor com os indices para ordenação.
}
