% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_polinomial_dic.R
\name{reg_polinomial_dic}
\alias{reg_polinomial_dic}
\title{Regressão polinomial (DIC)}
\usage{
reg_polinomial_dic(dados, x, y, bloco = NULL, grau = NULL, alpha = 0.05)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, linhas, colunas e, caso haja, blocos}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos, caso haja}

\item{grau}{grau maximo a ser considerado no modelo}

\item{alpha}{Nivel de significancia a ser utilizado, padrao 0.05}
}
\value{
Lista que contém:
\itemize{
\item \code{formulas} Lista com as formulas de todos os possíveis modelos de regressão até o \code{grau} definido.
\item \code{formula_str}Lista com as strings das formulas de todos os modelos de ANOVA de todos os possíveis modelos de regressão até o \code{grau} definido.
\item \code{modelos} Lista com  os modelos de ANOVA de todos os possíveis modelos de regressão até o \code{grau} definido.
\item \code{falta_ajuste} Lista com  as faltas de ajuste de todos os possíveis modelos de regressão até o \code{grau} definido.
\item \code{modelo} Modelo com o melhor ajuste.
\item \code{grau} Grau do modelo com melhor ajuste.
\item \code{plot} Gráfico dos dados com o modelo sobreposto.
\item \code{coeficiente} Coeficientes do modelo.
\item \code{fun} Função do modelo.
\item \code{derivada} Derivada da função do modelo.
\item \code{raiz} Valor da raiz da função do modelo.
\item \code{pmet} Ponto de máxima eficiência técnica.
}
}
\description{
Regressões polinomiais de delineamentos inteiramente casualizados
}
\examples{

bloco_casualizado(4, 1, 5)$dados \%>\%
  reg_polinom("Trat", "resultado")
 
}
