% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soma_grupo}
\alias{soma_grupo}
\title{Soma por grupos}
\usage{
soma_grupo(x, grupo)
}
\arguments{
\item{x}{Vetor numérico dos valores que serão somados}

\item{grupo}{Vetor com fatores que agruparão os dados}
}
\value{
Vetor com as somas resumidas de cada grupo.
}
\description{
Faz a soma de um vetor por grupos, semelhante a \code{tapply(y, x, sum)}.
}
