% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_rbcd.R
\name{anova_rbcd}
\alias{anova_rbcd}
\title{anova_rbcd}
\usage{
anova_rbcd(dados, x, y, bloco)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, os blocos e os resultados a serem analisados}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos}
}
\value{
list
}
\description{
Calcula valores utilizados para uma tabela de ANOVA de um delineamento em blocos completamente casualizados
}
