% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{bloco_incompleto}
\alias{bloco_incompleto}
\title{Blocos incompletos}
\usage{
bloco_incompleto(ntrat, nbloco, nrep = NULL)
}
\arguments{
\item{ntrat}{Numero de tratamentos a serem simulados}

\item{nbloco}{Numero de blocos a serem simulados}

\item{nrep}{Numero de repeticoes a serem simuladas, caso haja, padrao = NULL}
}
\value{
list
}
\description{
Simula um delineamento em blocos incompletos
}
