% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparacoes_lsqd.R
\name{comparacoes_lsqd}
\alias{comparacoes_lsqd}
\title{Comparacoes (Quadrados Latinos)}
\usage{
comparacoes_lsqd(dados, x, y, linha, coluna, alpha = 0.05, group = FALSE)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{alpha}{Nivel de significancia a ser utilizado, padrao 0.05}

\item{group}{Valor lógico, onde quando verdadeiro, fará a análise de comparacoes multiplas por grupos}

\item{bloco}{String com o nome da coluna dos blocos, manter NULL caso não hajam blocos}
}
\value{
comparacao
}
\description{
Análises múltiplas para ANOVA de delineamentos em quadrados latinos
}
