% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{estimando_ausentes}
\alias{estimando_ausentes}
\title{Estimando ausentes bibd}
\usage{
estimando_ausentes(dados, x, y, bloco)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, com valores faltantes, e blocos}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos}
}
\value{
data.frame
}
\description{
Estima os valores ausentes de um delineamento incompleto balanceado
}
