% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{grafico}
\alias{grafico}
\title{Grafico}
\usage{
grafico(dados, x, y, bloco = NULL, legenda = NULL)
}
\arguments{
\item{dados}{Data frame com colunas separadas para os tratamentos, resultados, e caso haja, blocos}

\item{x}{String com o nome da coluna dos tratamentos}

\item{y}{String com o nome da coluna dos resultados}

\item{bloco}{String com o nome da coluna dos blocos, manter NULL caso não hajam blocos}

\item{legenda}{String com a posição da legenda ("top", "bottom", "left", "right"), padrao = NULL}
}
\value{
ggplot
}
\description{
Funcao base para graficos de comparacoes multiplas
}
\details{
Funcoes de graficos
}
