% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.comparacao_lsqd}
\alias{plot.comparacao_lsqd}
\title{Grafico de comparacoes (Quadrados Latinos)}
\usage{
\method{plot}{comparacao_lsqd}(comparacao, ...)
}
\arguments{
\item{comparacao}{Objeto de comparacao_lsqd resultado da funcao \code{comparacao_lsqd}}
}
\value{
Ggplot com comparações
}
\description{
Graficos de comparacoes multiplas de delineamentos em quadrados latinos
}
\examples{

comparacao <- quadrados_latinos(4)$dados \%>\%
  comparacoes_lsqd("tratamento", "resultado", "linha", "coluna")
  
plot(comparacao)

}
