% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrados_latinos.R
\name{quadrados_latinos}
\alias{quadrados_latinos}
\title{Simula delineamento de quadrados latinos}
\usage{
quadrados_latinos(n, fun = rnorm)
}
\arguments{
\item{n}{Número de linhas (ou colunas)}

\item{fun}{Função que gere valores aleatórios cujo primeiro argumento é a quantidade de valores a serem gerados, Padrão = rnorm()}
}
\value{
Banco de dados com colunas dos tratamentos, linhas e colunas
}
\description{
Simula um delineamento em quadrados latinos a partir de um banco de dados
}
\examples{

quadrados_latinos(4)

}
