% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrados_latinos.R
\name{quadrados_latinos}
\alias{quadrados_latinos}
\title{Simula delineamento de quadrados latinos}
\usage{
quadrados_latinos(dados, tratamentos)
}
\arguments{
\item{dados}{Data frame com os resultados}

\item{tratamentos}{Vetor com fatores dos tratamentos}
}
\value{
Banco de dados com colunas dos tratamentos, linhas e colunas
}
\description{
Simula um delineamento em quadrados latinos a partir de um banco de dados
}
\examples{

quadrados_latinos(tibble(y = rnorm(16)), 1:4)

}
